Split out Links tests into own file
